+++ packages/orq-rc/src/models/operations/listagents.ts
@@ -1075,7 +1075,7 @@ export const ListAgentsKnowledgeBases$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01K6B3TVW2B7MYQDP7BF0RD33X"),
+  id: z.string().default("01K6CEZNNSWJ5SWXDHF8AMH8SR"),
   knowledge_id: z.string(),
   configuration: z.union([
     z.lazy(() =>
@@ -1106,7 +1106,7 @@ export const ListAgentsKnowledgeBases$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListAgentsKnowledgeBases
 > = z.object({
-  id: z.string().default("01K6B3TVW2B7MYQDP7BF0RD33X"),
+  id: z.string().default("01K6CEZNNSWJ5SWXDHF8AMH8SR"),
   knowledgeId: z.string(),
   configuration: z.union([
     z.lazy(() =>