+++ packages/orq-rc/RELEASES.md
@@ -3949,3 +3949,13 @@ Based on:
 - [typescript v3.13.0-rc.47] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.47] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.47 - packages/orq-rc
+
+## 2025-09-30 04:46:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.631.5 (2.716.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.48] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.48] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.48 - packages/orq-rc